refactor(socket): extract wait_for_socket to SocketCommon for reuse
Moved the static wait_for_socket() function from Socket-iov.c to
SocketCommon as SocketCommon_wait_for_fd() to enable reuse across
socket modules.

Changes:
- Added SocketCommon_wait_for_fd() declaration to SocketCommon.h
- Implemented SocketCommon_wait_for_fd() in SocketCommon.c
- Updated Socket-iov.c to use SocketCommon_wait_for_fd()
- Removed static wait_for_socket() function

The function provides a reusable mechanism for waiting on file
descriptors with timeout support, handling EINTR automatically.

Closes #564
